requirements.txt
pyspark==3.5.8
pymongo==4.6.1
pyyaml==6.0.1